requirements.txt
numpy
numba
matplotlib
pandas
jupyter
torch 
torchvision
comet-ml
opencv-python
tqdm
gym[atari]
pyyaml
comet_ml